Crate rome_cli

rome_cli

Rome’s main binary distribution, exposes the command line interface defined in this crate, and the language server interface defined in rome_lsp and used by the rome VSCode extension

Logs

When the server is run in daemon mode, it will output logs to a file created in a rome-logs directory inside the system temporary directory. The log file will be rotated on a hourly basis.

Module

This is where the main CLI session starts. The module is responsible to parse commands and arguments, redirect the execution of the commands and execute the traversal of directory and files, based on the command that were passed.

Structs

Arguments

An arguments parser.

CliSession

Global context for an execution of the CLI

FormatterReport

FormatterReportFileDetail

FormatterReportSummary

Report

ReportDiagnostic

Information computed from a diagnostic

ReportDiff

Information computed from a diff result

SocketTransport

Implementation of WorkspaceTransport for types implementing [AsyncRead] and [AsyncWrite]

Enums

ReportErrorKind

ReportKind

Termination

Error message returned by the CLI when it aborts with an error

Functions

open_transport

Tries to open a connection to a running daemon instance, returning a WorkspaceTransport instance if the socket is currently active

setup_panic_handler

Installs a global panic handler to show a user-friendly error message in case the CLI panics